Ingest storage file shares
