+++ src/azure/resource-manager/ResourceManagerClient.test.ts
@@ -1,23 +1,21 @@
-import { createTestLogger } from "@jupiterone/jupiter-managed-integration-sdk";
-
 import { VirtualMachine } from "@azure/arm-compute/esm/models";
+import { MySQLManagementModels } from "@azure/arm-mysql";
 import {
   NetworkInterface,
   NetworkSecurityGroup,
   PublicIPAddress,
   VirtualNetwork,
 } from "@azure/arm-network/esm/models";
-import { BlobContainer, StorageAccount } from "@azure/arm-storage/esm/models";
+import {
+  Database as SQLDatabase,
+  Server as SQLServer,
+} from "@azure/arm-sql/esm/models";
+import { createTestLogger } from "@jupiterone/jupiter-managed-integration-sdk";
 import { Polly } from "@pollyjs/core";
 
 import polly from "../../../test/helpers/polly";
 import config from "../../../test/integrationInstanceConfig";
 import { ResourceManagerClient } from "./";
-import {
-  Server as SQLServer,
-  Database as SQLDatabase,
-} from "@azure/arm-sql/esm/models";
-import { MySQLManagementModels } from "@azure/arm-mysql";
 
 let p: Polly;
 
@@ -29,7 +27,7 @@ test("client accessToken fetched once and used across resources", async () => {
   let requests = 0;
 
   p = polly(__dirname, "accessTokenCaching");
-  p.server.any().on("request", _req => {
+  p.server.any().on("request", (_req) => {
     requests++;
   });
 
@@ -58,7 +56,7 @@ describe("iterateNetworkInterfaces", () => {
     const client = new ResourceManagerClient(config, createTestLogger());
 
     const vms: NetworkInterface[] = [];
-    await client.iterateNetworkInterfaces(e => {
+    await client.iterateNetworkInterfaces((e) => {
       vms.push(e);
     });
 
@@ -81,7 +79,7 @@ describe("iterateNetworkSecurityGroups", () => {
     const client = new ResourceManagerClient(config, createTestLogger());
 
     const securityGroups: NetworkSecurityGroup[] = [];
-    await client.iterateNetworkSecurityGroups(e => {
+    await client.iterateNetworkSecurityGroups((e) => {
       securityGroups.push(e);
     });
 
@@ -110,7 +108,7 @@ describe("iteratePublicIPAddresses", () => {
     const client = new ResourceManagerClient(config, createTestLogger());
 
     const addresses: PublicIPAddress[] = [];
-    await client.iteratePublicIPAddresses(e => {
+    await client.iteratePublicIPAddresses((e) => {
       addresses.push(e);
     });
 
@@ -133,7 +131,7 @@ describe("iterateMySqlServers", () => {
     const client = new ResourceManagerClient(config, createTestLogger());
 
     const resources: MySQLManagementModels.Server[] = [];
-    await client.iterateMySqlServers(e => {
+    await client.iterateMySqlServers((e) => {
       resources.push(e);
     });
 
@@ -187,7 +185,7 @@ describe("iterateMySqlDatabases", () => {
     const client = new ResourceManagerClient(config, createTestLogger());
 
     const resources: MySQLManagementModels.Database[] = [];
-    await client.iterateMySqlDatabases(server, e => {
+    await client.iterateMySqlDatabases(server, (e) => {
       resources.push(e);
     });
 
@@ -251,7 +249,7 @@ describe("iterateSqlServers", () => {
     const client = new ResourceManagerClient(config, createTestLogger());
 
     const resources: SQLServer[] = [];
-    await client.iterateSqlServers(e => {
+    await client.iterateSqlServers((e) => {
       resources.push(e);
     });
 
@@ -288,7 +286,7 @@ describe("iterateSqlDatabases", () => {
     const client = new ResourceManagerClient(config, createTestLogger());
 
     const resources: SQLDatabase[] = [];
-    await client.iterateSqlDatabases(server, e => {
+    await client.iterateSqlDatabases(server, (e) => {
       resources.push(e);
     });
 
@@ -336,7 +334,7 @@ describe("iterateVirtualMachines", () => {
     const client = new ResourceManagerClient(config, createTestLogger());
 
     const vms: VirtualMachine[] = [];
-    await client.iterateVirtualMachines(e => {
+    await client.iterateVirtualMachines((e) => {
       vms.push(e);
     });
 
@@ -359,7 +357,7 @@ describe("iterateVirtualNetworks", () => {
     const client = new ResourceManagerClient(config, createTestLogger());
 
     const vms: VirtualNetwork[] = [];
-    await client.iterateVirtualNetworks(e => {
+    await client.iterateVirtualNetworks((e) => {
       vms.push(e);
     });
 
@@ -373,104 +371,4 @@ describe("iterateVirtualNetworks", () => {
       }),
     ]);
   });
-});
-
-describe("iterateStorageAccounts", () => {
-  test("all", async () => {
-    p = polly(__dirname, "iterateStorageAccounts");
-
-    const client = new ResourceManagerClient(config, createTestLogger());
-
-    const sa: StorageAccount[] = [];
-    await client.iterateStorageAccounts(e => {
-      sa.push(e);
-    });
-
-    expect(sa).toEqual([
-      expect.objectContaining({
-        id: expect.any(String),
-        name: "j1dev",
-        tags: expect.objectContaining({
-          environment: "j1dev",
-        }),
-      }),
-    ]);
-  });
-});
-
-describe("iterateStorageBlobContainers", () => {
-  test("all", async () => {
-    p = polly(__dirname, "iterateStorageBlobContainers");
-
-    const client = new ResourceManagerClient(config, createTestLogger());
-
-    const containers: BlobContainer[] = [];
-    await client.iterateStorageBlobContainers(
-      {
-        id:
-          "/subscriptions/dccea45f-7035-4a17-8731-1fd46aaa74a0/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/j1dev",
-        name: "j1dev",
-      } as StorageAccount,
-      e => {
-        containers.push(e);
-      },
-    );
-
-    expect(containers).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          id: expect.stringMatching(
-            /Microsoft\.Storage\/storageAccounts\/j1dev\/blobServices\/default\/containers\/bootdiagnostics-j1dev-/,
-          ),
-          name: expect.stringMatching(/bootdiagnostics-j1dev-/),
-        }),
-        expect.objectContaining({
-          id:
-            "/subscriptions/dccea45f-7035-4a17-8731-1fd46aaa74a0/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/j1dev/blobServices/default/containers/j1dev",
-          name: "j1dev",
-        }),
-      ]),
-    );
-  });
-
-  // skipped because jest.useFakeTimers() wouldn't work
-  test.skip(
-    "retry",
-    async () => {
-      // jest.useFakeTimers();
-
-      p = polly(__dirname, "iterateStorageBlobContainersRetry", {
-        recordFailedRequests: true,
-      });
-
-      const client = new ResourceManagerClient(config, createTestLogger());
-
-      let containers: BlobContainer[] = [];
-
-      // Get past the 100/5 min limit, be sure we get more than 100 just over 5 minutes
-      for (let index = 0; index < 103; index++) {
-        containers = [];
-        await client.iterateStorageBlobContainers(
-          {
-            id:
-              "/subscriptions/dccea45f-7035-4a17-8731-1fd46aaa74a0/resourceGroups/j1dev/providers/Microsoft.Storage/storageAccounts/j1dev",
-            name: "j1dev",
-          } as StorageAccount,
-          e => {
-            containers.push(e);
-          },
-        );
-      }
-
-      expect(containers).toEqual([
-        expect.objectContaining({
-          type: "Microsoft.Storage/storageAccounts/blobServices/containers",
-        }),
-        expect.objectContaining({
-          type: "Microsoft.Storage/storageAccounts/blobServices/containers",
-        }),
-      ]);
-    },
-    1000 * 1000, // allow this test to run long enough to hit the limit
-  );
 });
