+++ terraform/storage.tf
@@ -15,6 +15,12 @@ resource "azurerm_storage_container" "j1dev" {
   storage_account_name = azurerm_storage_account.j1dev.name
 }
 
+resource "azurerm_storage_share" "j1dev" {
+  name                 = "j1dev"
+  storage_account_name = azurerm_storage_account.j1dev.name
+  quota                = 1
+}
+
 resource "azurerm_sql_server" "j1dev" {
   count = var.azurerm_storage_sql_servers
